# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chatbot",
        "description": "API para la gestión de usuarios, chats y mensajes de un chatbot. Documentación en español.",
        "version": "1"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión de usuario",
                "description": "Autentica a un usuario y genera un token de acceso.",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secreto123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserDTO"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    },
                    "403": {
                        "description": "Cuenta no activa"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener información del usuario autenticado",
                "description": "Devuelve los datos del usuario autenticado y sus tokens.",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserDTO"
                                        },
                                        "tokens": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Revoca el token actual y cierra la sesión del usuario autenticado.",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Refrescar token de acceso",
                "description": "Genera un nuevo token de acceso y revoca el anterior.",
                "operationId": "30c20c52f5f72828d8c531fd7c9ea3e7",
                "responses": {
                    "200": {
                        "description": "Token refrescado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/chats": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "Listar chats del usuario autenticado",
                "description": "Obtiene una lista paginada de chats del usuario autenticado.",
                "operationId": "a80c3dfb6bf3308ef3246df062fee81c",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de chats por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de chats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "chats": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ChatDTO"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "Crear un nuevo chat",
                "description": "Crea un nuevo chat para el usuario autenticado.",
                "operationId": "b49d6a86afd5c3ed67550138804d654b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Estado del chat: active, closed, etc.",
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chat creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chat created successfully"
                                        },
                                        "chat": {
                                            "$ref": "#/components/schemas/ChatDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chats/{chatId}": {
            "put": {
                "tags": [
                    "Chats"
                ],
                "summary": "Actualizar un chat",
                "description": "Actualiza un chat específico del usuario autenticado.",
                "operationId": "f9109fd90b5ae9eda03e0c89d455c7f2",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID del chat a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "Estado del chat: active, closed, etc.",
                                        "type": "string",
                                        "example": "closed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chat actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chat updated successfully"
                                        },
                                        "chat": {
                                            "$ref": "#/components/schemas/ChatDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Chat no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chats"
                ],
                "summary": "Eliminar un chat",
                "description": "Elimina permanentemente un chat específico del usuario autenticado.",
                "operationId": "493a18ac0a32b66bf30d40e6c975e01e",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID del chat a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chat deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Chat no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chats/{chatId}/close": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "Cerrar un chat",
                "description": "Cierra un chat específico del usuario autenticado.",
                "operationId": "ef0e1be39bf10cd5fe680253a00887ca",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID del chat a cerrar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat cerrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chat closed successfully"
                                        },
                                        "chat": {
                                            "$ref": "#/components/schemas/ChatDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Chat no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Listar mensajes paginados",
                "description": "Obtiene una lista paginada de mensajes.",
                "operationId": "4216f3eedb7be911361f1b91030253a8",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de mensajes por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de mensajes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MessageDTO"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Enviar mensaje a un chat y obtener respuesta de IA",
                "description": "Envía un mensaje a un chat y recibe la respuesta del bot.",
                "operationId": "c85a9171473a13f151a8672235e01559",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "chat_id",
                                    "message"
                                ],
                                "properties": {
                                    "chat_id": {
                                        "description": "ID del chat",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "description": "Mensaje a enviar",
                                        "type": "string",
                                        "example": "Hola, ¿cómo estás?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mensaje enviado y respuesta recibida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_message": {
                                            "$ref": "#/components/schemas/MessageDTO"
                                        },
                                        "bot_response": {
                                            "type": "string",
                                            "example": "¡Hola! ¿En qué puedo ayudarte hoy?"
                                        },
                                        "chat_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-25T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/messages/chat/{chatId}": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Obtener mensajes por chat",
                "description": "Devuelve todos los mensajes de un chat específico.",
                "operationId": "2fb06712e46d63f0ed5b3d519aa53990",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID del chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensajes del chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MessageDTO"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Chat no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChatDTO": {
                "title": "Chat",
                "description": "DTO que representa un chat entre usuario y bot.",
                "required": [
                    "user_id",
                    "status"
                ],
                "properties": {
                    "user_id": {
                        "description": "ID del usuario asociado al chat",
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "description": "Estado del chat: active, closed, etc.",
                        "type": "string",
                        "example": "active"
                    },
                    "last_message_at": {
                        "description": "Fecha y hora del último mensaje en el chat",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-25T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
            "UserDTO": {
                "title": "Usuario",
                "description": "DTO que representa a un usuario del sistema.",
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "password": {
                        "description": "Contraseña del usuario (solo para registro o actualización)",
                        "type": "string",
                        "format": "password",
                        "example": "secreto123"
                    },
                    "role": {
                        "description": "Rol asignado al usuario",
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación mediante token Bearer. Ingrese 'Bearer {token}' en el encabezado Authorization.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}